Removed a few TODOs (literally!)

+++ src/preprocess.h
@@ -60,8 +60,13 @@ void rm_file_tables_destroy(RmFileTables *list);
  */
 bool rm_file_tables_insert(RmSession *session, RmFile *file);
 
-// TODO
-long rm_pp_cmp_orig_criteria_impl(
+/**
+ * @brief Compare certain attributes (listed below) of files 
+ *        in order to find out which file is the original.
+ *
+ * Returns: 
+ */
+int rm_pp_cmp_orig_criteria_impl(
     RmSession *session,
     time_t mtime_a, time_t mtime_b,
     const char *basename_a, const char *basename_b,
